+++ t1_initialization/t1_initialization/t1_initialization.vcxproj
@@ -179,6 +179,7 @@
     </FxCompile>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="Box.cpp" />
     <ClCompile Include="camera.cpp" />
     <ClCompile Include="DDSTextureLoader.cpp" />
     <ClCompile Include="input.cpp" />
@@ -190,8 +191,10 @@
     <ClCompile Include="texture.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="Box.h" />
     <ClInclude Include="camera.h" />
     <ClInclude Include="DDSTextureLoader.h" />
+    <ClInclude Include="def.h" />
     <ClInclude Include="input.h" />
     <ClInclude Include="plane.h" />
     <ClInclude Include="renderer.h" />
