+++ t1_initialization/t1_initialization/t1_initialization.vcxproj.filters
@@ -37,9 +37,6 @@
     <Filter Include="Scene">
       <UniqueIdentifier>{9cfeebf5-3960-4805-a58e-fd943daeccbb}</UniqueIdentifier>
     </Filter>
-    <Filter Include="Scene\Shaders">
-      <UniqueIdentifier>{25f8e279-1f20-4f40-bcaf-0a753724ad3a}</UniqueIdentifier>
-    </Filter>
     <Filter Include="Skybox\Shaders">
       <UniqueIdentifier>{9fcf3e78-edd4-424f-8fce-e9084a398031}</UniqueIdentifier>
     </Filter>
@@ -51,6 +48,12 @@
     </Filter>
     <Filter Include="Scene\Plane\Shaders">
       <UniqueIdentifier>{53d8f888-0b61-4452-89b7-4d0515e34c92}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Scene\Box">
+      <UniqueIdentifier>{c55094f5-967f-437a-850c-68bce334578e}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Scene\Box\Shaders">
+      <UniqueIdentifier>{25f8e279-1f20-4f40-bcaf-0a753724ad3a}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
@@ -78,6 +81,9 @@
     </ClCompile>
     <ClCompile Include="plane.cpp">
       <Filter>Scene\Plane</Filter>
+    </ClCompile>
+    <ClCompile Include="Box.cpp">
+      <Filter>Scene\Box</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -111,6 +117,10 @@
     <ClInclude Include="plane.h">
       <Filter>Scene\Plane</Filter>
     </ClInclude>
+    <ClInclude Include="Box.h">
+      <Filter>Scene\Box</Filter>
+    </ClInclude>
+    <ClInclude Include="def.h" />
   </ItemGroup>
   <ItemGroup>
     <FxCompile Include="skybox_PS.hlsl">
@@ -120,10 +130,10 @@
       <Filter>Skybox\Shaders</Filter>
     </FxCompile>
     <FxCompile Include="t2_PS.hlsl">
-      <Filter>Scene\Shaders</Filter>
+      <Filter>Scene\Box\Shaders</Filter>
     </FxCompile>
     <FxCompile Include="t2_VS.hlsl">
-      <Filter>Scene\Shaders</Filter>
+      <Filter>Scene\Box\Shaders</Filter>
     </FxCompile>
     <FxCompile Include="transparent_PS.hlsl">
       <Filter>Scene\Plane\Shaders</Filter>
